+++ s/tools/lit_listener.ts
@@ -1,9 +1,9 @@
 
-export interface LitListener<E extends Event> extends AddEventListenerOptions {
+export interface LitListener<E extends Event = Event> extends AddEventListenerOptions {
 	handleEvent(e: E): void
 }
 
-export function litListener<E extends Event>(o: LitListener<E>) {
+export function litListener<E extends Event = Event>(o: LitListener<E>) {
 	return o
 }